Split an app into multiple pages
