+++ doccano_mini/models.py
@@ -1,0 +1,8 @@
+# https://platform.openai.com/docs/models/gpt-3-5
+AVAILABLE_MODELS = (
+    "gpt-3.5-turbo",
+    "gpt-3.5-turbo-0301",
+    "text-davinci-003",
+    "text-davinci-002",
+    "code-davinci-002",
+)
