+++ doccano_mini/prompts.py
@@ -3,8 +3,6 @@ from typing import List
 from langchain.prompts.few_shot import FewShotPromptTemplate
 from langchain.prompts.prompt import PromptTemplate
 
-from .tasks import TaskType
-
 
 def make_classification_prompt(examples: List[dict]) -> FewShotPromptTemplate:
     unique_labels = set([example["label"] for example in examples])
@@ -39,11 +37,3 @@ def make_task_free_prompt(examples: List[dict]) -> FewShotPromptTemplate:
         input_variables=columns[:-1],
     )
     return prompt
-
-
-def select_prompt_maker(task: TaskType):
-    if task == TaskType.TEXT_CLASSIFICATION.value:
-        return make_classification_prompt
-    elif task == TaskType.TASK_FREE.value:
-        return make_task_free_prompt
-    raise ValueError(f"Task {task} is not supported.")
